Activity cancelled event not received for Call Activity
Potential issue in that I do not see ACTIVITY_CANCELLED event when the
parent process is deleted.  I see the ACTIVITY_CANCELLED event for the
user task within the external subprocess and the PROCESS_CANCELLED for the
external subprocess, but not the ACTIVITY_CANCELLED for the call activity.
Included test recreates the issue