add candidateAssignedToOthers flag to TaskInfoQuery interface
This extra flag has been implemented in order to allow querying without assignee logic.
When setting the flag when querying, one can still request tasks or historic tasks regardless of the set assignee.
The behaviour without this flag is that a task that is assigned can never be returned when searching for candidateUsers.