Add injection of the process engine to trigger an intialization of the ProcessEngine
If the ProcessEngine is not injected it can happen that the method is called before the ProcessEngine is initialized
which leads to the other engines not being initialized yet.

`@ConditionalOnBean` only checks that a bean definition is present - and does not create the bean right away to avoid early initialization issues.