Fix documentation property for LDAP initialization user
The LDAP initialization user is set by the Flowable IDM app only and
it is app specific. That's the reason why it has the prefix
`flowable.idm` and not `flowable.common.app.idm`.

As reference, it is set in the following method:
org.flowable.ui.idm.conf.Bootstrapper.initializeDefaultPrivileges

Both calls are refering to flowable.idm.app.admin.user-id.