Add search for LDAP group by group id (resolves #763)
Currently it's not possible to see group details or assign a privilege
to a group in case you have more than one groups. Reason therefore is
that for the LDAPQueryBuilder the group id is ignored.

New behavior is that there is a new attribute to search groups by
the specified group-id. Therefore a new search query is introduced.
For the sample application the user needs to configure
 `flowable.idm.ldap.query.group-by-id`.