Cleanup Flowable Spring Boot Security Auto configuration
Flowable should only configure the UserDetailsService.
Spring security performs a default configuration of the authentication provider using a provider UserDetailsService and PasswordEncoder
through the InitializeUserDetailsBeanManagerConfigurer.
In case a global AuthenticationProvider is configured then Spring Security uses that through the InitializeAuthenticationProviderBeanManagerConfigurer.
Doing this allows easier configuration of Flowable Spring Boot applications with a custom UserDetailsService

Add the possibility to use the Spring Security 5 DelegatingPasswordEncoder through a property