Add common HasVariableTypes interface and make sure that VariableTypes are shared across all engines
VariableTypes were not shared during serialization time (writing to DB), but only during deserialization (reading from DB)

When writing to DB each engine uses their own VariableTypes to perform the persistence. However, when reading from DB
the IbatisVariableTypeHandler is used and this handler is shared across all engines (through the shared DefaultSqlSessionFactory).
This meant that when an engine custom variable type (JPA in ProcessEngine) when the AppEngine is in the lead then
writing JPA variable works, but reading that variable was not working as the JPA variable was not registered in the AppEngine.

This commit makes sure that the VariableTypes are shared both during write and read of variables.