Use transient variables for the WebServiceActivityBehavior and remove CustomObjectType
The implementation of the CustomObjectType was like a transient variable as it was only using the cached value.
However, a variable was still being persisted in the database in which all fields were null, this variable was / is useless.

The ItemInstance and MessageInstance are handled as transient variables, before they were persisted in the database
via the CustomObjectType, and they had no persisted value, they only had cached value.