Clarify DMN Deployment definition example
In the existing example, only the `addClasspathResource` method was used. This method  uses the classpath string to create the InputStream internally and to name the resource. In the other methods available, like addInputStream, you must provide in two parameters the resource name and the resource variable to get the bytes from the definition. In these cases, if you don't set the resource name with a .dmn suffix, the deployment does not recognize the resource as a DMN definition. In the existing example, the  name of the deployment was set without a suffix and that was also misleading.