Revert "Update Mockito to version 2.24.0 (#1571)" (#1588)
This reverts commit 3ff0874f6715f065d233723f7de3f5889e905480.

The reason is that the tests in the different UI apps are now failing due to conflicting versions of ByteBuddy. We are importing the spring-boot-dependencies in the pom and the apps are using that. This leads to having Mockito 2.24.0 and ByteBuddy 1.7.11 (defined in the spring-boot-dependencies). Mockito 2.24.0 needs ByteBuddy 1.9.7.

Tests are also failing on Java 12 EA when using Mockito 2.24.0 with ByteBuddy 1.9.7. Failure is: "Mockito cannot mock this class: interface java.util.Map. Mockito can only mock non-private & non-final classes."