Explicitly disable the LDAP health endpoint in For the Flowable UI Idm tests
Before 2.1.0 the LdapHealthIndicator was not created due to the missing LdapOperations.
However since 2.1 the LdapOperations bean creation was moved into the LdapAutoConfiguration (see https://github.com/spring-projects/spring-boot/pull/13136)
which lead to the LdapHealthIndicator being created and thus the FlowableIdmApplicationSecurityTest was failing due to the ldap user and password being null.
See  https://github.com/spring-projects/spring-boot/pull/17861 for the problem in Spring Boot