Allow MyBatis mapping file to be customized.
Our use of flowable requires some default mybatis type handlers to be overriden. Based upon the mybatis documentation the only way to do this is by specifying the overides in the mapping file as documented here:
https://mybatis.org/mybatis-3/configuration.html#typeHandlers

This PR therefore introduces new methods to allow for the location of the mybatis mapping file to be chosen.

Change-Id: Iff9af696a950c5d79b0ea4c2fa96582a5c3b1842