Use @Configuration(proxyBeanMethods = false) where possible
proxyBeanMethods is a new attribute on the Spring Configuration (since 5.2),
its purpose is to tell Spring not the create CGLib proxies for the configuration classes.
Even though it was added in 5.2, having it doesn't make applications using older Spring not work.
It is still possible to use Flowable with Spring Boot 1.5