Minor changes in Spanish translation + script (#2163)
* Improved spanish translation

Modules translated:
  - flowable-ui-admin
  - flowable-ui-idm
  - flowable-ui-modeler
  - flowable-ui-tash

* Minor changes in Spanish translation + script

Minor changes in Spanish translation.

The files were edited to match line-by-line with the English version.

A script was added to detect changes in the structure of the
localization files. This script allows to identify new or deleted keys
according to the English version, to keep a line-by-line matching.