Avoid NPE when criterion doesn't have a sentry
An exit criterion doesn't have a sentry for non blocking tasks.
The reason for this is that the CMMN spec does not allow criteria on non blocking tasks.
However, you can write them in the CMMN XML and the CmmnXmlConverter would skip this criteria
(i.e. the task would have no exit criteria, thus the sentry will never be set in the Exit Criterion)