Add support for system channel events
Add support for sending system events outbound. The default implementation is an in memory implementation that invokes the event registry event consumers.
The BPMN and CMMN send event tasks should be able to send system events by using the event registry
Refactor the EventInstance and remove the not needed EventModel and Channel models from it
Merge the event payload and correlation parameters into the event payload
Make the OutboundEventChannelAdapter and OutboundEventProcessingPipeline generic interfaces so that they can have different implementations then just String