Use assertj fluent style in flowable-rest module. (#2330)
* Use assertj fluent style in flowable-rest module.
The module had a mix of assertion styles.  This is part 2 for this module.
This finishes off the `org.flowable.rest.service.api.history` package.

* Use assertj fluent style in flowable-rest module.
The module had a mix of assertion styles.  This is part 2 for this module.
This finishes off the `org.flowable.rest.service.api.history` package.