Use assertj fluent style in flowable-rest module. (#2333)
* Use assertj fluent style in flowable-rest module.
The module had a mix of assertion styles.  This is part 4 for this module.
This is for the `org.flowable.rest.service.api.repository` package.
Once the first part is merged I will come back and remove the private method that I included in some of the files as it is part of the base class.

* Use assertj fluent style in flowable-rest module. (#2330)

* Use assertj fluent style in flowable-rest module.
The module had a mix of assertion styles.  This is part 2 for this module.
This finishes off the `org.flowable.rest.service.api.history` package.

* Use assertj fluent style in flowable-rest module.
The module had a mix of assertion styles.  This is part 2 for this module.
This finishes off the `org.flowable.rest.service.api.history` package.

* Use assertj fluent style in flowable-rest module. (#2332)

The module had a mix of assertion styles.  This is part 3 for this module.
This finishes off the `org.flowable.rest.service.api.identity` and `org.flowable.rest.service.api.management` packages.

* Use assertj fluent style in flowable-rest module.
The module had a mix of assertion styles.  This is part 4 for this module.
This is for the `org.flowable.rest.service.api.repository` package.
Once the first part is merged I will come back and remove the private method that I included in some of the files as it is part of the base class.

* Use assertj fluent style in flowable-rest module.
The module had a mix of assertion styles.  This is part 4 for this module.
This is for the `org.flowable.rest.service.api.repository` package.
Move duplicated `getISODateStringWithTZ()` method into base class.