Add check to prevent NPE.
While converting a test file to assertj I found an example of a cut-n-paste that didn't test the method that the author thought it did.  So when I changed the method name to the correct method name and reran the test a NPE was generated instead of the expected `FlowableIllegalArgumentException`.   Looking in the source file I noticed that all other "ignore case" methods checked for `null` before trying `value.toLowerCase()` except in the case being tested.  This fixes that error.