Use assertj fluent style in flowable-engine module. (#2419)
* Use assertj fluent style in flowable-engine module.
In the `org.flowable.engine.test.bpmn.event.timer` package there were 4 files with mixed assertion styles so they were converted to a single format. At the same time the remaining files in the package were updated.   This is part 1 of 2.

* Use assertj fluent style in flowable-engine module.
In the org.flowable.engine.test.bpmn.event.timer package there were 4 files with mixed assertion styles so they were converted to a single format. At the same time the remaining files in the package were updated. This is part 2 of 2.