Do not fail the entire transaction when using grouping of history json data when a single json processing fails
Collect the failed data that failed and schedule new job instead of rollbacking the entire transaction and removing the progress for the successful processing.
This makes sure that when there are 10 json nodes in one history job, failing of 1 will only have missing data for that one and not for all the data in the job.