Make sure that deploying Kafka, JMS and Rabbit inbound channels does not interfere with the default Spring Registry
When a Rabbit / JMS listener container is started immediately it will mean that the endpoint registry itself is running.
This means that listeners registered via @RabbitListener and @JmsListener will not be started and thus won't receive messages.

There is no need for doing anything specific in the Kafka implementation since the KafkaListenerEndpointRegistry stores its running state independent of the underlying containers.