Improve parallel multi instance leave to avoid optimistic locking exceptions
Store number of completed and active instances for a parallel multi instance behaviour as a custom variable type that doesn't need updates and can be computed.
Doing this remove the optimistic locking exceptions when multiple parallel executions complete in the same time.

Perform leave logic check in an exclusive asynchronous job in order to avoid the lock on the parent scope when parallel multi instance executions complete.
This allows executing non exclusive asynchronous parallel multi instance in parallel without an exception being thrown.