Enhance Tomcat EL util class with one from the eclipse/el-ri + do not coerce value if the value is null
Adapt the Util#findWrapper method in order to support better ambiguous method overloads for the BeanELResolver.
This supports looking up in interfaces in order to pick the method that would have been picked by the Java compiler.
e.g. TestAmbiguousMethodSingleExecutionParameterBean from the tests does not work with the Tomcat Util implementation, but it does with the eclipse/el-ri one

The not coercing of the value when it is null is done in order to keep the previous behaviour and to avoid backwards incompatibility.
The tests InvokeXXXMethodWithNullParameter are used for this