Update logging levels in job service. (#3008)
Update the logging level for some exception cases from error to warn in the job service.

Cause:
These errors occur in our k8s environment when the primary database server fails over to a secondary. This failover can take up to a few seconds to occur resulting in the database being inaccessible during that time.

Justification:
Often an error message in the logs is an indicator that a system administrator must perform some investigation and work to resolve the issue. However, here that is not the case since the the logic will retry. To reduce the amount of information that a system administrator needs to look through, we would like to update the logging here to warn to reflect that an immediate action isn't required. If the warning continues, then a system administrator should investigate.