Decouple the channel body deserialization from the InboundEventDeserializer
Make sure that all previous APIs are binary compatible with the previous version.

With this commit we make sure that the event deserialization is decoupled from the different channel implementations.
It introduces a new InboundEvent that makes it possible to  provide the headers based on the channel type.
It is a step in a direction towards providing a way to extract custom channel message specific elements into the event.
e.g. Kafka topic, partition, offset, etc.