Task/refactor to variable container (#3411)
* Refactor ScriptEngines to VariableContainer instead of VariableScope

Change the APIs to consume VariableContainer instead of VariableScope in scripting. VariableContainer is easier to handle.

* Make changes backwards compatible

Remove getVariables() from VariableContainer again, as there are implementation already, which do not rely on a map and cannot easily provide this method.
Instead check for variableScope in ScriptBindings (in methods which are not really used by javax.script it seems.
Ensures backwards compatible runtime behavior.