Change the lock approach for the event registry unique instance
Change the approach for locking for the check of a unique instance.
It is possible to have multiple threads trying to start independent unique instances.
When the lock is done on the event registry then only one of the instances (instead of the 2) will be created.

With the new approach we are using our lock manager to acquire a lock specific to the unique values for the instance that will be created.
This means that both of the instances from the scenario above will be created