#3452 Support throwing BPMNError from executionListeners
BPMNErrors thrown from executionListeners can now be handled in error Event Sub-Process or boundary events from sub process invocations in the model.
ErrorBoundary events on the activity where the executionListeners are defined are not invoked. Instead, a parent event handler is looked up, when an executionListener throws an BPMNError and invoked, if found.
This is to be consistent with the execution behavior of timer boundary events and signal boundary events.