fix: NPE fix when signal not found from model
In some rare cases the BpmnModel can change between the call to
"containsSignalId" and the call to get the signal, causing an NPE
if the signal was no longer available on the model.

Due to the way containsSignalId is implemented, the call to it
is entirely unnecessary in these cases.