feat: in-memory data managers for various entity types
A new module providing high performance non-sql in memory data managers for various entity types.

The implementation is based on Concurrent maps serving as the storage layer of the entities. This
removes the chance of locking exceptions entirely and increases the performance of process execution
significantly (2-5x faster compared to H2 or hsqldb even for simple processes), but the managers
do not support (full) transactions. They implement a simple rollback / commit strategy where things
like deadletter jobs work, and items are removed from memory on rollback, but many parts of true
transactions, like rolling back updated objects, are not yet implemented.

Any data manager that does not have a no-sql implementation will keep using the existing database
based datamanagers. When running the non-sql in-memory datamanagers, it makes most sense to also
run an in-memory database (eg. H2 or hsqldb) instead of a real database. The database is used for
things like process deployments and definitions, byte arrays and other entities that either have
limited performance improvements from memory data managers, or are complicated to implement due
to references to table data on the database level.

The in-memory data managers are for runtime processes only. It is intended that an executor using
them has history set excplicitly to NONE - if history is enabled the benefits of the in-memory
data managers vanish because each process still needs to do database operations to store history.