Improve the change detection manager to properly register latest definitions
Change the DefaultInboundChannelModelCacheManager take the channel definition version into consideration
when deciding whether a channel needs to be registered or not.
Only channels with versions higher than the one that is currently registered should be considered for registration.

Expose the currently registered channels through the InboundChannelModelCacheManager and use it in the DefaultEventRegistryChangeDetectionManager to make sure that we only unregister deleted channels