Fix for repeating process task in CMMN creating a new user task on deletion of the case instance
The reason is that the deletion of the process instance causes the RuntimeInstanceCallback to be called which leads to the planning of a termination of the associated plan item instance of the process task. As the plan item instance was already terminates (due to the case instance being deleted), this should be a no-op.