Increase the interval time for checking if there are jobs in the InstantTimeTimerEventTest
The InstantTimeTimerEventTest is a test which test timer jobs.
This means that we can have a scenario when the wait returns, but there are still jobs.
The scenario (due to race conditions is the following).

We start with a timer job and start the async executor

* Test Thread: Check async job list is empty
* Async Thread: Move timer job to async job
* Test Thread: Check timer job list is empty

Due to race conditions the test thread thinks that there are no more jobs and stops waiting and continues to the next check.