Use dedicated linked list for storing the EvaluateCriteriaOperation
The EvaluateCriteriaOperation(s) are special types of operations
that should be executed only after the regular operations are executed.
The previous implementation was iterating the operations to find the first EvaluateCriteriaOperation
and insert the regular operation before that one.
However, when there are a lot of operations iterating the list is costly.
Therefore, using a different list avoids the costly iterations