HeadersPayloadExtractor should support case insensitive event headers (#4138)
When a custom event implementation is using case insensitive headers then our `HeadersPayloadExtractor` was not properly working due to the fact that we were checking if a header with a specific name is configured on our side. However, we should do the opposite. Check if there is a header with the configured name in the event headers.